add numpy as dependency, drop PY2 testing
