+++ plugins/thumbnails/plugin.cmake
@@ -1,5 +1,4 @@
 add_python_test(thumbnail PLUGIN thumbnails)
-add_python_test(thumbnailPy2 PLUGIN thumbnails PY2_ONLY)
 add_python_style_test(python_static_analysis_thumbnails
                       "${PROJECT_SOURCE_DIR}/plugins/thumbnails/server")
 add_python_style_test(python_static_analysis_thumbnails_tests
