+++ plugins/thumbnails/plugin_tests/thumbnail_test.py
@@ -174,6 +174,110 @@ class ThumbnailsTestCase(base.TestCase):
         self.model('folder').remove(self.publicFolder)
         self.assertEqual(self.model('file').load(thumbnailId), None)
 
+    def testDicomThumbnailCreation(self):
+        path = os.path.join(ROOT_DIR, 'plugins', 'thumbnails', 'plugin_tests', 'data',
+                            'sample_dicom.dcm')
+        with open(path, 'rb') as file:
+            data = file.read()
+
+        # Upload the Girder logo to the admin's public folder
+        resp = self.request(
+            path='/file', method='POST', user=self.admin, params={
+                'parentType': 'folder',
+                'parentId': self.publicFolder['_id'],
+                'name': 'test.dcm',
+                'size': len(data)
+            })
+        self.assertStatusOk(resp)
+        uploadId = resp.json['_id']
+
+        fields = [('offset', 0), ('uploadId', uploadId)]
+        files = [('chunk', 'test.dcm', data)]
+        resp = self.multipartRequest(
+            path='/file/chunk', fields=fields, files=files, user=self.admin)
+        self.assertStatusOk(resp)
+        self.assertIn('itemId', resp.json)
+        fileId = resp.json['_id']
+
+        params = {
+            'fileId': fileId,
+            'width': 64,
+            'attachToId': str(self.admin['_id']),
+            'attachToType': 'user'
+        }
+
+        # We shouldn't be able to add thumbnails without write access to the
+        # target resource.
+        resp = self.request(
+            path='/thumbnail', method='POST', user=self.user, params=params)
+        self.assertStatus(resp, 403)
+
+        # Should complain if we don't pass a width or a height
+        del params['width']
+        params['attachToId'] = str(self.user['_id'])
+
+        resp = self.request(
+            path='/thumbnail', method='POST', user=self.user, params=params)
+        self.assertStatus(resp, 400)
+        self.assertEqual(resp.json['message'], 'You must specify a valid width,'
+                         ' height, or both.')
+
+        # Set a width, we should now correctly have a thumbnail
+        params['width'] = 64
+        resp = self.request(
+            path='/thumbnail', method='POST', user=self.user, params=params)
+        self.assertStatusOk(resp)
+        job = resp.json
+
+        from girder.plugins.jobs.constants import JobStatus
+        self.assertEqual(job['status'], JobStatus.SUCCESS)
+
+        self.user = self.model('user').load(self.user['_id'], force=True)
+        self.assertEqual(len(self.user['_thumbnails']), 1)
+        thumbnailId = self.user['_thumbnails'][0]
+
+        resp = self.request('/file/%s/download' % str(thumbnailId),
+                            isJson=False)
+        data = self.getBody(resp, text=False)
+        image = Image.open(six.BytesIO(data))
+        self.assertEqual(image.size, (64, 64))
+
+        # Delete the thumbnail, it should be removed from the user thumb list
+        resp = self.request('/file/%s' % str(thumbnailId), method='DELETE',
+                            user=self.user)
+        self.assertStatusOk(resp)
+
+        self.assertEqual(self.model('file').load(thumbnailId), None)
+        self.user = self.model('user').load(self.user['_id'], force=True)
+        self.assertEqual(len(self.user['_thumbnails']), 0)
+
+        # Attach a thumbnail to the admin's public folder
+        resp = self.request(
+            path='/thumbnail', method='POST', user=self.admin, params={
+                'width': 64,
+                'height': 32,
+                'crop': True,
+                'attachToId': str(self.publicFolder['_id']),
+                'attachToType': 'folder',
+                'fileId': fileId
+            })
+        self.assertStatusOk(resp)
+        self.publicFolder = self.model('folder').load(
+            self.publicFolder['_id'], force=True)
+        self.assertEqual(len(self.publicFolder['_thumbnails']), 1)
+
+        thumbnailId = self.publicFolder['_thumbnails'][0]
+
+        resp = self.request('/file/%s/download' % str(thumbnailId),
+                            isJson=False)
+        data = self.getBody(resp, text=False)
+        image = Image.open(six.BytesIO(data))
+        self.assertEqual(image.size, (64, 32))
+
+        # Deleting the public folder should delete the thumbnail as well
+        self.model('folder').remove(self.publicFolder)
+        self.assertEqual(self.model('file').load(thumbnailId), None)
+
     def testCreateThumbnailOverride(self):
         def override(event):
             # Override thumbnail creation -- just grab the first 4 bytes
