+++ plugins/thumbnails/requirements.txt
@@ -1,2 +1,3 @@
 Pillow==3.1.0
 pydicom==0.9.9
+numpy==1.11.1
